rewrite character update

+++ Engine/MagicManager.cs
@@ -152,7 +152,7 @@ namespace Engine
 
         private static MagicSprite GetMoveMagicSprite(Character user, Magic magic, Vector2 origin, Vector2 destination, bool destroyOnEnd, float speedRatio = 1f)
         {
-            var speed = (int)(Globals.MagicBasespeed * magic.Speed * speedRatio);
+            var speed = Globals.MagicBasespeed * magic.Speed * speedRatio;
             return new MagicSprite(
                 magic,
                 user,
@@ -164,7 +164,7 @@ namespace Engine
 
         private static MagicSprite GetMoveMagicSpriteOnDirection(Character user, Magic magic, Vector2 origin, Vector2 direction, bool destroyOnEnd, float speedRatio = 1f)
         {
-            var speed = (int)(Globals.MagicBasespeed * magic.Speed * speedRatio);
+            var speed = Globals.MagicBasespeed * magic.Speed * speedRatio;
             return new MagicSprite(
                 magic,
                 user,
